Add files via upload
Hospitals, medical devices, clinical trials, outsourcing, telemedicine, medical tourism, health insurance, and medical equipment are all part of the healthcare industry. The industry is expanding at a rapid pace as a result of improved coverage, services, and increased spending by both public and private players. And with the growth of AI in all the sectors, it has also become a helping hand in the healthcare sector with the help of cognitive computing. Cognitive computing is transforming healthcare delivery around the world, and all these frameworks use automated models to recreate human perspectives. On a global scale, these computing systems have been critical in deciphering the massive volume of healthcare data. This paper is going to put light on how important cognitive computing has become for getting better outcome from the health care sector, be it medical research personalized medial solution of illness of each patient or predicting the future possibilities of health issues in patients. There many questions which this paper is going to address in the field of cognitive computing in healthcare. This paper is also focused on the achievement that technology has made in past decades such as how Computer vision is helping visually challenged to make lives a little easy, also how IBM is using cognitive computing in making treatment system more personalized.